Resolve using repositories declaired in all projects (#2)
Previously resolution occured on the project using the repositories
it is configured with. For multi-module projects this may be just a
subset of the repositories used. Commonly the root project has no
dependencies and therefore no repositories. Some child projects may
add a repository for a dependency, again creating a mismatch.

Now all additional repositories are collected, temporarily added to
the project, and removed after resolution completes. The repositories
used are logged at `info` level.

Tested using the example project.