Remove guard clause due to bad method invocation.
The no-arg `getFirstLevelModuleDependencies()` method is only availabe
on the default implementation and is not defined on the interface for
LenientConfiguration. The anonymous implementation returned from when
any empty dependency set is used lacks this helper method, resulting
in a runtime failure. The interface method is now invoked with a
filter to collect everything.