Refactor plugin to be configuration aware
Previously the dependencies were resolved in a detached configuration,
which had a few bad side effects. One is that the same dependency with
different versions would be resolved to the latest and be shown once.
The other is that the configuration's resolutionStrategy and compoent
rules were not honored.

Now every configuration is resolved (safely by cloning) and the report
an aggregate of the results. This lets us decide if we should auto
include all repositories found, which may sometimes cause misbehaviors.

This is a major refactoring to transition the code from a scripting
style to a Java style. This is mostly needed for code clarity, as the
former style made sense when the plugin was untyped. The contribution
to being type-safe was a nice addition, but made the code appear
awkward.

The ability to define custom resolution strategies allows filtering
the report to disallow release candidates or other exclusion policies.