Parallelize and cache project url resolution
The artifact query is not cached by Gradle, making it a slow sequential
call to multiple http endpoints (repositories). A dependency may be in
multiple configurations, causing repeated lookups. This change helps in
a JDK6 compatible fashion by caching the lookup and using the GPars pool
to query in parallel. Unfortunately, I think the query is synchronous
within Gradle, so the pool does not assist much.