Add flag to disable gradle update check (fixes #241)
The task property `checkForGradleUpdate` is defaulted to `true`. When set to
`false` no check is performed, the plain text report, does not include the
statement, and the json/xml report includes an enable flag, no version, and
the reason.

Due to URLConnection defaulting to infinite timeout that can cause a hung build,
the maximum connection and read timeouts are set to 15 seconds as a failsafe.