Remove incompatible parallelization (fixes #255)
Unfortunately the GPars trick is not safe, causing deadlocks with other
plugins. The safest solution is to remove and eventually use newer
Gradle apis that expose some parallel support. The core performance
issue, though, is that Gradle does not resolve dependencies in parallel
so doing project & configurations was a marginal gain.