Remove  deblibs-gradle-plugin from the list (#580)
Hey @ben-manes. Unfortunately we decided to stop support for https://github.com/hellofresh/deblibs-gradle-plugin. 

This PR removes a reference to it, so we can avoid redirecting people to deprecated plugin which will be really hard to integrate. If at any point we will decide to update it and put it back we would let you know. 

Thank you very much!