Fix inherited collection to not be always empty
This fixes a mistake in PR #721 which used allDependencies here instead of dependencies, causing the inherited set always be empty, as a result of subtracting allDependencies from it.

This also prevents inherited kotlin deps from being included in the latest set, to avoid issues with having the original and modifies dependencies both added to the copy.